+++ markdown-link-check.sh
@@ -1,5 +1,0 @@
-#|/bin/bash
-
-markdown-link-check -q docs
-markdown-link-check -q README.md
-